#142 Refactored roles helper methods + improved db_roles view
